Added functional component and added txData to use right data for gas fee
