+++ ui/pages/send/gas-display/index.js
@@ -1,1 +1,1 @@
-export { default } from './gas-display.component';
+export { default } from './gas-display';
