+++ ui/pages/send/send.test.js
@@ -42,6 +42,10 @@ jest.mock('react-router-dom', () => {
   };
 });
 
+setBackgroundConnection({
+  getGasFeeTimeEstimate: jest.fn(),
+});
+
 jest.mock('ethers', () => {
   const originalModule = jest.requireActual('ethers');
   return {
@@ -64,6 +68,14 @@ const baseStore = {
   },
   history: { mostRecentOverviewPage: 'activity' },
   metamask: {
+    unapprovedTxs: {
+      1: {
+        id: 1,
+        txParams: {
+          value: 'oldTxValue',
+        },
+      },
+    },
     gasEstimateType: GAS_ESTIMATE_TYPES.LEGACY,
     gasFeeEstimates: {
       low: '0',
